// src/types/objectState.ts
export enum ObjectState {
  OBJSTA_NONE = 0x00000000,

  OBJSTA_STAND = 0x00000001,
  OBJSTA_STAND2 = 0x00000002,
  OBJSTA_SIT = 0x00000003,
  OBJSTA_FMOVE = 0x00000004,
  OBJSTA_BMOVE = 0x00000005,
  OBJSTA_LEFT = 0x00000006,
  OBJSTA_RIGHT = 0x00000007,
  OBJSTA_PICKUP = 0x00000008,

  OBJSTA_LMOVE = 0x00000009,
  OBJSTA_RMOVE = 0x0000000a,
  OBJSTA_STOP_RUN = 0x0000000b,

  OBJSTA_MOVE_ALL = 0x000000ff,
  OBJSTA_NOT_STAND = OBJSTA_MOVE_ALL & ~OBJSTA_STAND,

  // === Turn Left/Right ===
  OBJSTA_LTURN = 0x00000100,
  OBJSTA_RTURN = 0x00000200,
  OBJSTA_TURN_ALL = 0x00000300,

  // === Look up/down
  OBJSTA_LOOKUP = 0x00000400,
  OBJSTA_LOOKDOWN = 0x00000800,
  OBJSTA_LOOK_ALL = 0x00000c00,

  // === Jump ===
  OBJSTA_FJUMP_READY = 0x00001000,
  OBJSTA_FJUMP = 0x00002000,
  OBJSTA_FFALL = 0x00003000,
  OBJSTA_FLAND = 0x00004000,
  OBJSTA_SJUMP1 = 0x00005000,
  OBJSTA_SJUMP2 = 0x00006000,
  OBJSTA_SJUMP3 = 0x00007000,
  OBJSTA_SJUMP4 = 0x00008000,
  OBJSTA_BJUMP_READY = 0x00009000,
  OBJSTA_BJUMP = 0x0000a000,
  OBJSTA_BFALL = 0x0000b000,
  OBJSTA_BLAND = 0x0000c000,
  OBJSTA_JUMP_ALL = 0x0000f000,

  // === Basic attack ===
  OBJSTA_ATK1 = 0x00010000,
  OBJSTA_ATK2 = 0x00020000,
  OBJSTA_ATK3 = 0x00030000,
  OBJSTA_ATK4 = 0x00040000,

  // === Wand attack ===
  OBJSTA_ATK_MAGIC1 = 0x00050000,
  OBJSTA_ATK_RANGE1 = 0x00060000,
  OBJSTA_RANGE3 = 0x00070000,
  OBJSTA_RANGE4 = 0x00080000,

  // === Melee skills ===
  OBJSTA_ATK_MELEESKILL = 0x00090000,
  OBJSTA_ATK_RANGESKILL = 0x000a0000,
  OBJSTA_ATK_CASTING1 = 0x000b0000,
  OBJSTA_ATK_CASTING2 = 0x000c0000,
  OBJSTA_ATK_MAGICSKILL = 0x000d0000,
  OBJSTA_SP_ATK1 = 0x00100000,
  OBJSTA_SP_ATK2 = 0x00200000,
  OBJSTA_ATK_ALL = 0x00ff0000,

  // === Damages ===
  OBJSTA_DMG = 0x01000000,
  OBJSTA_DMG_RESERVED1 = 0x04000000,
  OBJSTA_DMG_RESERVED2 = 0x05000000,
  OBJSTA_DMG_FLY = 0x02000000,
  OBJSTA_DMG_DIE = 0x03000000,
  OBJSTA_DMG_LIVE = 0x06000000,
  OBJSTA_DMG_FLY_ALL = 0x02000000,

  // === Dead ===
  OBJSTA_DEAD = 0x08000000,
  OBJSTA_DISAPPEAR = 0x09000000,
  OBJSTA_RESURRECTION = 0x0c000000,
  OBJSTA_DEAD_RESERVED1 = 0x0d000000,
  OBJSTA_DIE_ALL = 0x08000000,
  OBJSTA_DMG_ALL = 0x0f000000,

  // === Other ===
  OBJSTA_COLLECT = 0x10000000,
  OBJSTA_APPEAR = 0x20000000,
  OBJSTA_APPEAR2 = 0x30000000,
  OBJSTA_STUN = 0x40000000,
  OBJSTA_ACTION_ALL = 0xf0000000,

  OBJSTA_ALL = OBJSTA_MOVE_ALL |
    OBJSTA_TURN_ALL |
    OBJSTA_LOOK_ALL |
    OBJSTA_JUMP_ALL |
    OBJSTA_ATK_ALL |
    OBJSTA_DMG_ALL |
    OBJSTA_ACTION_ALL,
}